moved tiles definition handling from HitReco to Subsystem
